ZKsync OS REVM (#2)
ZKsync OS variant of REVM. Most of the code is just imports from REVM
and wrappers that implement REVM traits. The big things to review:
- `precompiles` — precompiles should behave exactly the same as on
ZKsync OS, and if a precompile is present on ZKsync OS, then it must be
present in this REVM and vice versa.
- `hooks` — same as precompiles. All ZKsync OS hooks should be
implemented here as well. Pay attention to:
- `gas`. How much used by precompiles in a positive testcase. Out of Gas
handling.
    - How the precompile should behave in `static` mode.
    - What would happen if hook is called in `delegatecall`.
- `tx_types` - all transaction types should behave exactly the same way
as on ZKsync OS. For EIP-1559 transactions, this means correctly
charging fees and not burning the base fee. For pre-155 txs, don’t check
the chain ID. For L1 → L2 transactions (priority and upgrade types),
don’t perform validation, and don’t bump the nonce when validating the
transaction.
- `ZKsyncHandler` - this is the main component. Most of the custom logic
is implemented here. It handles transaction validation, refunds, and
state change reverts. Pay special attention to `gas_used_override` and
`force_failing`. I recommend reading the vanilla REVM. Unfortunately, I
couldn’t find good documentation for this component.
-
https://docs.rs/revm-handler/latest/revm_handler/handler/trait.Handler.html
    - https://reth.rs/docs/reth/revm/revm/handler/trait.Handler.html

---------

Co-authored-by: Daniyar Itegulov <ditegulov@gmail.com>